+++ packages/react-widgets/src/widgets/PieWidget.js
@@ -20,7 +20,7 @@ import useWidgetLoadingState from './useWidgetLoadingState';
  * @param  {string} props.column - Name of the data source's column to get the data from.
  * @param  {string} [props.operationColumn] - Name of the data source's column to operate with. If not defined it will default to the one defined in `column`.
  * @param  {string} props.operation - Operation to apply to the operationColumn. Must be one of those defined in `AggregationTypes` object.
- * @param  {Object} [props.colorsPerCategory] - (optional) Dictionary with RGB colors per category, expressed as { 'Cat 1': [255, 0, 255] }.
+ * @param  {Object} [props.colorsPerCategory] - (optional) Dictionary with hex colors per category, expressed as { 'Cat 1': '#fabada' }.
  * @param  {formatterCallback} [props.formatter] - Function to format the value that appears in the tooltip.
  * @param  {formatterCallback} [props.tooltipFormatter] - Function to return the HTML of the tooltip.
  * @param  {boolean} [props.viewportFilter=true] - Defines whether filter by the viewport or not.
@@ -75,10 +75,10 @@ function PieWidget({
       })
         .then((data) => {
           setCategoryData(data);
-          if (data && colorsPerCategory) {
+          if (colorsPerCategory) {
             const colors = data.map((category) => {
               const categoryColor = colorsPerCategory[category.name];
-              const othersColor = [200, 200, 200];
+              const othersColor = '#777';
               return categoryColor || colorsPerCategory['Others'] || othersColor;
             });
             setCategoryColors(colors);
@@ -91,9 +91,7 @@ function PieWidget({
         .finally(() => setIsLoading(false));
     } else {
       setCategoryData([]);
-      if (colorsPerCategory) {
-        setCategoryColors([]);
-      }
+      setCategoryColors([]);
     }
 
     return function cleanup() {